Refactor rename package streak -> streaks
