+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile/presentation/ProfileFeature.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 import org.hyperskill.app.analytic.domain.model.AnalyticEvent
 import org.hyperskill.app.core.domain.url.HyperskillUrlPath
 import org.hyperskill.app.profile.domain.model.Profile
-import org.hyperskill.app.streak.domain.model.Streak
+import org.hyperskill.app.streaks.domain.model.Streak
 
 interface ProfileFeature {
     sealed interface State {
