--- shared/src/commonMain/kotlin/org/hyperskill/app/streak/domain/repository/StreaksRepository.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.streak.domain.repository
-
-import org.hyperskill.app.streak.domain.model.Streak
-
-interface StreaksRepository {
-    suspend fun getStreaks(userId: Long): Result<List<Streak>>
-}
