--- shared/src/commonMain/kotlin/org/hyperskill/app/streak/injection/StreaksDataComponent.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.streak.injection
-
-import org.hyperskill.app.streak.domain.interactor.StreaksInteractor
-
-interface StreaksDataComponent {
-    val streaksInteractor: StreaksInteractor
-}
