+++ shared/src/commonMain/kotlin/org/hyperskill/app/streaks/data/source/StreaksRemoteDataSource.kt
@@ -1,6 +1,6 @@
-package org.hyperskill.app.streak.data.source
+package org.hyperskill.app.streaks.data.source
 
-import org.hyperskill.app.streak.domain.model.Streak
+import org.hyperskill.app.streaks.domain.model.Streak
 
 interface StreaksRemoteDataSource {
     suspend fun getStreaks(userId: Long): Result<List<Streak>>
