+++ shared/src/commonMain/kotlin/org/hyperskill/app/streaks/domain/repository/StreaksRepository.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.streaks.domain.repository
+
+import org.hyperskill.app.streaks.domain.model.Streak
+
+interface StreaksRepository {
+    suspend fun getStreaks(userId: Long): Result<List<Streak>>
+}
