+++ shared/src/commonMain/kotlin/org/hyperskill/app/streaks/injection/StreaksDataComponent.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.streaks.injection
+
+import org.hyperskill.app.streaks.domain.interactor.StreaksInteractor
+
+interface StreaksDataComponent {
+    val streaksInteractor: StreaksInteractor
+}
