+++ shared/src/commonMain/kotlin/org/hyperskill/app/streaks/injection/StreaksDataComponentImpl.kt
@@ -1,11 +1,11 @@
-package org.hyperskill.app.streak.injection
+package org.hyperskill.app.streaks.injection
 
 import org.hyperskill.app.core.injection.AppGraph
-import org.hyperskill.app.streak.data.repository.StreaksRepositoryImpl
-import org.hyperskill.app.streak.data.source.StreaksRemoteDataSource
-import org.hyperskill.app.streak.domain.interactor.StreaksInteractor
-import org.hyperskill.app.streak.domain.repository.StreaksRepository
-import org.hyperskill.app.streak.remote.StreaksRemoteDataSourceImpl
+import org.hyperskill.app.streaks.data.repository.StreaksRepositoryImpl
+import org.hyperskill.app.streaks.data.source.StreaksRemoteDataSource
+import org.hyperskill.app.streaks.domain.interactor.StreaksInteractor
+import org.hyperskill.app.streaks.domain.repository.StreaksRepository
+import org.hyperskill.app.streaks.remote.StreaksRemoteDataSourceImpl
 
 class StreaksDataComponentImpl(appGraph: AppGraph) : StreaksDataComponent {
     private val streaksRemoteDataSource: StreaksRemoteDataSource =
