+++ shared/src/commonMain/kotlin/org/hyperskill/app/streaks/remote/StreaksRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.streak.remote
+package org.hyperskill.app.streaks.remote
 
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
@@ -6,9 +6,9 @@ import io.ktor.client.request.get
 import io.ktor.client.request.parameter
 import io.ktor.http.ContentType
 import io.ktor.http.contentType
-import org.hyperskill.app.streak.data.source.StreaksRemoteDataSource
-import org.hyperskill.app.streak.domain.model.Streak
-import org.hyperskill.app.streak.remote.model.StreaksResponse
+import org.hyperskill.app.streaks.data.source.StreaksRemoteDataSource
+import org.hyperskill.app.streaks.domain.model.Streak
+import org.hyperskill.app.streaks.remote.model.StreaksResponse
 
 class StreaksRemoteDataSourceImpl(
     private val httpClient: HttpClient
